Added BaseApplication
Refatored code to move common code to application class
Could not get the downloading of the specialtags to work properly. Only half of the xml data was being read. It could have something to do with an undefined content-length on the xml being returned and thus the client connection has no way to know how much data is being sent unless it keeps on reading. The http connection class, okhttp, just reads a specified amount and then terminates.

So, what has been done is the response in a web browser has been downloaded and the XML response is saved in raw format for the app to use locally. This should be adequate for now until an explanation is determined as to why the downloading fails.